+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Introduces the median polish function `eda_pol`.
 * Introduces the QQ and Tukey mean-difference plot `eda.qq`.
+* Introduces the density plot function `eda.dens`.
 * Adds re-expression parameters to `eda_lm` via the parameters `px` and `py`.
 * Adds `sd` labels to SD dashed lines in `eda_lm`.
 * `eda_lm` will now output `lm` intercept and slope.